[update] moved tests to __test__ folder
