+++ src/__tests__/Events.test.js
@@ -1,18 +1,18 @@
 import React from "react";
-import Spacecrafts from "./Spacecrafts";
+import Events from "../components/Events";
 import { render } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 
 let getByTestId;
 
 beforeEach(() => {
-    const component = render(<Spacecrafts/>);
+    const component = render(<Events/>);
     getByTestId = component.getByTestId;
 })
 
 test("header renders with correct text", () => {
-    const headerElement = getByTestId("spacecraft-header");
+    const headerElement = getByTestId("events-header");
 
-    expect(headerElement.textContent).toBe("Spacecrafts");
+    expect(headerElement.textContent).toBe("Events");
 })
