+++ src/__tests__/Gallery.test.js
@@ -1,18 +1,18 @@
 import React from "react";
-import Locations from "./Locations";
+import Gallery from "../components/Gallery";
 import { render } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 
 let getByTestId;
 
 beforeEach(() => {
-    const component = render(<Locations/>);
+    const component = render(<Gallery/>);
     getByTestId = component.getByTestId;
 })
 
 test("header renders with correct text", () => {
-    const headerElement = getByTestId("locations-header");
+    const headerElement = getByTestId("gallery-header");
 
-    expect(headerElement.textContent).toBe("Locations");
+    expect(headerElement.textContent).toBe("Gallery");
 })
