+++ src/__tests__/Locations.test.js
@@ -1,18 +1,18 @@
 import React from "react";
-import Gallery from "./Gallery";
+import Locations from "../components/Locations";
 import { render } from "@testing-library/react";
 import "@testing-library/jest-dom/extend-expect";
 
 let getByTestId;
 
 beforeEach(() => {
-    const component = render(<Gallery/>);
+    const component = render(<Locations/>);
     getByTestId = component.getByTestId;
 })
 
 test("header renders with correct text", () => {
-    const headerElement = getByTestId("gallery-header");
+    const headerElement = getByTestId("locations-header");
 
-    expect(headerElement.textContent).toBe("Gallery");
+    expect(headerElement.textContent).toBe("Locations");
 })
